HACK: Ugly hack to make split working
Split currently doesn't support splits that happens in subelements. This needs
to be fixed correctly. In the meantime as I need it right now, I commit a
version that hacks around it in ugly way. The result is *NOT* well formed. But
luckily it works in some browsers (Firefox/Chrome).